google_alerts_enricher/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
trafilatura>=1.6.0
PyYAML>=6.0
